+++ src/frontend/apps/lti_site/apps/deposit/components/Dashboard/index.tsx
@@ -1,8 +1,8 @@
 import { Loader } from '@openfun/cunningham-react';
-import { Box, ThemeContext, ThemeType } from 'grommet';
+import { ThemeContext, ThemeType } from 'grommet';
 import { normalizeColor } from 'grommet/utils';
 import { theme } from 'lib-common';
-import { UploadManager, useCurrentResourceContext } from 'lib-components';
+import { Box, UploadManager, useCurrentResourceContext } from 'lib-components';
 import React, { Fragment } from 'react';
 import { FormattedMessage, defineMessages, useIntl } from 'react-intl';
 
@@ -105,7 +105,7 @@ const Dashboard = () => {
       <ThemeContext.Extend value={extendedTheme}>
         <Box
           align="center"
-          background="linear-gradient(-180deg, #dce9fa 0%, #EDF5FA 100%);"
+          background="linear-gradient(-180deg, #dce9fa 0%, #EDF5FA 100%)"
           pad="small"
         >
           {content}