+++ src/frontend/apps/lti_site/apps/deposit/components/SelectContent/SelectContentSection/index.tsx
@@ -1,8 +1,8 @@
 import { Button } from '@openfun/cunningham-react';
-import { Box, Grid } from 'grommet';
+import { Grid } from 'grommet';
 import { DocumentStore } from 'grommet-icons';
 import { Nullable } from 'lib-common';
-import { ContentCard, FileDepository, Text } from 'lib-components';
+import { Box, ContentCard, FileDepository, Text } from 'lib-components';
 import React from 'react';
 import { defineMessages, useIntl } from 'react-intl';