+++ src/frontend/apps/lti_site/components/DashboardDocument/DocumentPlayer/index.tsx
@@ -1,5 +1,5 @@
-import { Box } from 'grommet';
 import {
+  Box,
   Document,
   DocumentXapiStatement,
   useCurrentSession,