+++ src/frontend/apps/lti_site/components/SelectContent/SelectContentTabs/index.spec.tsx
@@ -197,10 +197,10 @@ describe('SelectContentTabs', () => {
       within(screen.getByLabelText('Select Webinar 1')).getByRole('img', {
         name: 'thumbnail',
       }),
-    ).toHaveStyleRule(
-      'background',
-      'url(https://example.com/default_thumbnail/144) no-repeat center / cover',
-    );
+    ).toHaveStyle({
+      background:
+        'url(https://example.com/default_thumbnail/144) no-repeat center / cover',
+    });
 
     screen.getByText('Webinar 1');
     screen.getByText('Webinar 2');
@@ -214,10 +214,10 @@ describe('SelectContentTabs', () => {
       within(screen.getByLabelText('Select Video 1')).getByRole('img', {
         name: 'thumbnail',
       }),
-    ).toHaveStyleRule(
-      'background',
-      'url(https://example.com/default_thumbnail/144) no-repeat center / cover',
-    );
+    ).toHaveStyle({
+      background:
+        'url(https://example.com/default_thumbnail/144) no-repeat center / cover',
+    });
 
     screen.getByText('Video 1');
     screen.getByText('Video 2');