+++ src/frontend/apps/lti_site/components/SelectContent/SelectContentTargetedResource/index.tsx
@@ -1,5 +1,5 @@
-import { Box } from 'grommet';
 import {
+  Box,
   BoxLoader,
   Document,
   Heading,